% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterObject.R
\name{combineTwoDimensionalTrace}
\alias{combineTwoDimensionalTrace}
\title{Combine Two-Dimensional Trace}
\usage{
combineTwoDimensionalTrace(trace1, trace2, max)
}
\arguments{
\item{trace1}{A trace read in from C++ in the form of a vector of vectors.
The last value of this trace should be equal to the first value of the second trace.
Once this function has executed, this is the trace that should be modified.}

\item{trace2}{A trace read in from C++ in the form of a vector of vectors.
The first value of this trace should be equal to the last value of the first trace.}

\item{max}{}
}
\value{
This function has no return value.
}
\description{
Combine Two-Dimensional Trace
}

