% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcObject.R
\name{convergence.test}
\alias{convergence.test}
\title{Convergence Test}
\usage{
convergence.test(trace, what, mixture = 1, n.samples = 10, frac1 = 0.1,
  frac2 = 0.5, plot = FALSE)
}
\arguments{
\item{trace}{}

\item{what}{}

\item{mixture}{}

\item{n.samples}{}

\item{frac1}{}

\item{frac2}{}

\item{plot}{}

\item{...}{}
}
\value{

}
\description{
\code{convergence.test}
}
\details{
\code{convergence.test}
}

