% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotParameterObject.R
\name{cspACF}
\alias{cspACF}
\title{Plots ACF for CSP traces}
\usage{
cspACF(parameter, csp = "Mutation", numMixtures = 1, samples = 500,
  lag.max = 40)
}
\arguments{
\item{parameter}{object of class Parameter}

\item{csp}{"Selection" or "Mutation", defaults to "Mutation"}

\item{numMixtures}{indicates the number of CSP mixtures used}

\item{samples}{number of samples to calculate ACF for,
Ex) if samples == 300 and hae 1000 samples total, ACF will be calculated from samples 700 to 1000}

\item{lag.max}{Maximum amount of lag to calculate ACF}
}
\description{
Plots ACF for CSP traces
}

