% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estimateElongationCost}
\alias{estimateElongationCost}
\title{estimateElongationCost}
\description{
Method of FONSEParameter class (access via parameter$estimateElongationCost(), where parameter is initialized with model="FONSE"). Enable estimation of the elongation cost parameter (a2), which is held fixed by default.
}
