% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estimateInitiationCost}
\alias{estimateInitiationCost}
\title{estimateInitiationCost}
\description{
Method of FONSEParameter class (access via parameter$estimateInitiationCost(), where parameter is initialized with model="FONSE"). Enable estimation of the initiation cost parameter (a1), which is held fixed by default.
}
