% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterObject.R
\name{extractBaseInfo}
\alias{extractBaseInfo}
\title{Extract Base Info}
\usage{
extractBaseInfo(parameter)
}
\arguments{
\item{parameter}{A Parameter object that corresponds to one of the model types.
Valid values are "ROC", "RFP", and "FONSE".}
}
\value{
varlist A list containing various attributes of the Parameter base class.
}
\description{
Extract Base Info
}

