% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fixAlpha}
\alias{fixAlpha}
\title{fixAlpha}
\description{
Method of PAParameter or PANSEParameter class (access via parameter$fixAlpha(), where parameter is initialized with model="PA" or "PANSE"). Fix the pausing shape parameter (alpha) at its current values for the remainder of the MCMC run.
}
