% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fixDOmega}
\alias{fixDOmega}
\title{fixDOmega}
\description{
Method of FONSEParameter class (access via parameter$fixDOmega(), where parameter is initialized with model="FONSE"). Fix the nonsense error coefficients (dOmega) at their current values for the remainder of the MCMC run.
}
