% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fixedElongationCost}
\alias{fixedElongationCost}
\title{fixedElongationCost}
\description{
Method of FONSEParameter class (access via parameter$fixedElongationCost(), where parameter is initialized with model="FONSE"). Fix the elongation cost parameter (a2) at its current value for the remainder of the MCMC run.
}
