% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fixedInitiationCost}
\alias{fixedInitiationCost}
\title{fixedInitiationCost}
\description{
Method of FONSEParameter class (access via parameter$fixedInitiationCost(), where parameter is initialized with model="FONSE"). Fix the initiation cost parameter (a1) at its current value for the remainder of the MCMC run.
}
