% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fixLambdaPrime}
\alias{fixLambdaPrime}
\title{fixLambdaPrime}
\description{
Method of PAParameter or PANSEParameter class (access via parameter$fixLambdaPrime(), where parameter is initialized with model="PA" or "PANSE"). Fix the elongation rate parameter (lambdaPrime) at its current values for the remainder of the MCMC run.
}
