% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fixNSERate}
\alias{fixNSERate}
\title{fixNSERate}
\description{
Method of PANSEParameter class (access via parameter$fixNSERate(), where parameter is initialized with model="PANSE"). Fix the per-codon nonsense error rates at their current values for the remainder of the MCMC run.
}
