% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fixSphi}
\alias{fixSphi}
\title{Fix the value of s_phi (standard deviation of lognormal for synthesis rates) at its current value}
\description{
Fix the value of s_phi (standard deviation of lognormal for synthesis rates) at its current value
}
