% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fixZ}
\alias{fixZ}
\title{fixZ}
\description{
Method of PANSEParameter class (access via parameter$fixZ(), where parameter is initialized with model="PANSE"). Fix the partition function Z at its current value for the remainder of the MCMC run.
}
