% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptiveScheme.R
\name{format.AdaptiveScheme}
\alias{format.AdaptiveScheme}
\title{Format an AdaptiveScheme as a one-line summary string.}
\usage{
\method{format}{AdaptiveScheme}(x, ...)
}
\arguments{
\item{x}{An AdaptiveScheme.}

\item{...}{Unused.}
}
\value{
Character vector of length 1 or 2.
}
\description{
Format an AdaptiveScheme as a one-line summary string.
}
