% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi.prior.spec.R
\name{format.prior_dist}
\alias{format.prior_dist}
\title{Format a prior_dist object as a string}
\usage{
\method{format}{prior_dist}(x, ...)
}
\arguments{
\item{x}{A \code{prior_dist} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A character string describing the distribution.
}
\description{
Format a prior_dist object as a string
}
