% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getEstimatedMixtureAssignmentForGene}
\alias{getEstimatedMixtureAssignmentForGene}
\title{Get estimated mixture assignment for gene}
\arguments{
\item{samples}{number of samples over which to calculate mixture assignment}

\item{geneIndex}{corresponding index of gene in genome. Should be a number between 1 and length(genome).}
}
\value{
returns value between 1 and n, where n is number of mixtures
}
\description{
Get estimated mixture assignment for gene
}
