% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getEstimatedMixtureAssignmentProbabilitiesForGene}
\alias{getEstimatedMixtureAssignmentProbabilitiesForGene}
\title{Get estimated mixture assignment probabilities for gene}
\arguments{
\item{samples}{number of samples over which to calculate mixture assignment probabilities}

\item{geneIndex}{corresponding index of gene in genome. Should be a number between 1 and length(genome).}
}
\value{
returns vector of probabilities representing mixture probabilities for gene
}
\description{
Get estimated mixture assignment probabilities for gene
}
