% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getGroupList}
\alias{getGroupList}
\title{Get amino acids (ROC, FONSE) or codons (PA, PANSE) for which parameters will be estimated}
\description{
Get amino acids (ROC, FONSE) or codons (PA, PANSE) for which parameters will be estimated
}
