% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getIterationsToAdapt}
\alias{getIterationsToAdapt}
\title{getIterationsToAdapt}
\value{
number of iterations to adapt
}
\description{
Method of MCMC class (access via mcmc$<function name>, where mcmc is an object initialized by initializeMCMCObject). Return number of iterations (iterations = samples * thinning) to allow proposal widths to adapt
}
