% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getLogPosteriorMean}
\alias{getLogPosteriorMean}
\title{Calculate the mean log posterior probability over the last n samples}
\arguments{
\item{samples}{postive value less than total length of the MCMC trace}
}
\description{
Calculate the mean log posterior probability over the last n samples
}
