% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getNoiseOffsetPosteriorMean}
\alias{getNoiseOffsetPosteriorMean}
\title{Calculate posterior mean of standard deviation parameter of lognormal describing distribution of synthesis rates}
\arguments{
\item{index}{mixture index to use. Should be number between 0 and n-1, where n is number of mixtures}

\item{samples}{number of samples over which to calculate posterior mean}
}
\value{
returns posterior mean for standard deviation of lognormal distribution of synthesis rates
}
\description{
Calculate posterior mean of standard deviation parameter of lognormal describing distribution of synthesis rates
}
