% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getStepsToAdapt}
\alias{getStepsToAdapt}
\title{Return number of iterations (total iterations = samples * thinning) to allow proposal widths to adapt}
\description{
Return number of iterations (total iterations = samples * thinning) to allow proposal widths to adapt
}
