% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getSynthesisRate}
\alias{getSynthesisRate}
\title{Get current synthesis rates for all genes and all mixtures}
\value{
2 by 2 vector of numeric values
}
\description{
Get current synthesis rates for all genes and all mixtures
}
