% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getSynthesisRatePosteriorMeanForGene}
\alias{getSynthesisRatePosteriorMeanForGene}
\title{Get posterior mean synthesis rate value for a gene}
\arguments{
\item{samples}{number of samples over which to calculate mean}

\item{geneIndex}{corresponding index of gene in genome for which posterior mean synthesis rate will be calculated. Should be a number between 1 and length(genome)}

\item{log_scale}{Calculate posterior mean on log scale}
}
\value{
posterior mean synthesis rate for gene
}
\description{
Get posterior mean synthesis rate value for a gene
}
