% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getThinning}
\alias{getThinning}
\title{Return thinning value, which is the number of iterations (total iterations = samples * thinning) not being kept}
\description{
Return thinning value, which is the number of iterations (total iterations = samples * thinning) not being kept
}
