% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterObject.R
\name{getTrace}
\alias{getTrace}
\title{extracts an object of traces from a parameter object.}
\usage{
getTrace(parameter)
}
\arguments{
\item{parameter}{A Parameter object that corresponds to one of the model types.}
}
\value{
trace Returns an object of type Trace extracted from the given parameter object
}
\description{
extracts an object of traces from a parameter object.
}
