% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getTraceObject}
\alias{getTraceObject}
\title{Get Trace object stored by a Parameter object. Useful for plotting certain parameter traces.}
\description{
Get Trace object stored by a Parameter object. Useful for plotting certain parameter traces.
}
