% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterObject.R
\name{initializeParameterObject}
\alias{initializeParameterObject}
\title{Initialize Parameter}
\usage{
initializeParameterObject(
  genome = NULL,
  sphi = NULL,
  num.mixtures = 1,
  gene.assignment = NULL,
  initial.expression.values = NULL,
  model = "ROC",
  split.serine = TRUE,
  mixture.definition = "allUnique",
  mixture.definition.matrix = NULL,
  init.with.restart.file = NULL,
  mutation.prior.mean = 0,
  mutation.prior.sd = 0.35,
  propose.by.prior = FALSE,
  selection.prior.mean = 0,
  selection.prior.sd = 100,
  init.csp.variance = 0.0025,
  init.sepsilon = 0.1,
  init.w.obs.phi = FALSE,
  init.by.random = FALSE,
  init.initiation.cost = 4,
  init.partition.function = 1,
  numElongationMixtures = 1,
  phi.prior = "lognormal",
  phi.prior.constraint = "mean",
  phi.prior.init = NULL,
  phi.prior.hyperparams = NULL,
  phi.mphi = NULL,
  phi.sphi = NULL
)
}
\arguments{
\item{genome}{An object of type Genome necessary for the initialization of the Parameter object.
The default value is NULL.}

\item{sphi}{Initial value(s) for sphi (log-space SD of the phi prior, equivalent
to \code{sdlog} in \code{\link{dlnorm}}). Numeric vector of length numMixtures, or
\code{NA} to estimate sphi with the default weakly informative prior (see
\code{phi.sphi}). Will be superseded by \code{phi.sphi} when that argument is
provided.}

\item{num.mixtures}{The number of mixtures elements for the underlying mixture distribution (numMixtures > 0).
The default value is 1.}

\item{gene.assignment}{A vector holding the initial mixture assignment for each gene. 
The vector length has to equal the number of genes in the genome.
Valid values for the vector range from 1 to numMixtures. 
It is possible but not advised to leave a mixture element empty.
The default Value is NULL.}

\item{initial.expression.values}{(Optional) A vector with intial phi values.
The length of the vector has to equal the number of genes in the Genome object and the order of the genes should match the order of the genes in the Genome.
The default value is NULL.}

\item{model}{Specifies the model used. Valid options are "ROC", "PA", "PANSE", or "FONSE".
The default model is "ROC".
ROC is described in Gilchrist et al. 2015.
PA, PANSE and FONSE are currently unpublished.}

\item{split.serine}{Whether serine should be considered as 
one or two amino acids when running the model.
TRUE and FALSE are the only valid values.
The default value for split.serine is TRUE.}

\item{mixture.definition}{A string describing how each mixture should
be treated with respect to mutation and selection.
Valid values consist of "allUnique", "mutationShared", and "selectionShared" for ROC/FONSE.
Valid values consist of  "allUnique", "elongationShared", and "nseShared" for PANSE.
The default value for mixture.definition is "allUnique".
See details for more information.}

\item{mixture.definition.matrix}{A matrix representation of how
the mutation and selection categories correspond to the mixtures.
The default value for mixture.definition.matrix is NULL.
If provided, the model will use the matrix to initialize the mutation and selection
categories instead of the definition listed directly above.
See details for more information.}

\item{init.with.restart.file}{File name containing information to reinitialize a 
previous Parameter object.
If given, all other arguments will be ignored.
The default value for init.with.restart.file is NULL.}

\item{mutation.prior.mean}{Controlling the mean of the normal prior on mutation paramters.
If passed in as single number (default is 0), this will be the mean value for all categories, for all codons. User may also
supply a vector with n * 40 values, where n is the number of mutation categories. Future versions will check the number of rows matches
the number of mutation categories definded by user.}

\item{mutation.prior.sd}{Controlling the standard deviation of the normal prior on the mutation parameters.
If passed in as single number (default is 0.35), this will be the standard deviation value for all categories, for all codons. User may also
supply a vector with n * 40 values, where n is the number of mutation categories. Future versions will check the number of rows matches
the number of mutation categories definded by user.}

\item{propose.by.prior}{Mutation bias parameters will be proposed based on the means and standard deviations set in mutation.prior.mean and mutation.prior.sd.  Argument is to be depricated and replaced by `propose.by.prior.mutaton` and `propose.by.prior.selection` created.}

\item{selection.prior.mean}{ROC only. Controlling the mean of the normal prior on selection paramters.
If passed in as single number (default is 0), this will be the mean value for all categories, for all codons. User may also
supply a vector with n * 40 values, where n is the number of selection categories. Future versions will check the number of rows matches
the number of selection categories definded by user.}

\item{selection.prior.sd}{ROC only. Controlling the standard deviation of the normal prior on the selection parameters.
If passed in as single number (default is 100 which effectively results in a flat prior), this will be the standard deviation value for all categories, for all codons. User may also
supply a vector with n * 40 values, where n is the number of selection categories. Future versions will check the number of rows matches
the number of selection categories definded by user.}

\item{init.csp.variance}{specifies the initial proposal width for codon specific parameter (default is 0.0025). 
The proposal width adapts during the runtime to reach a taget acceptance rate of ~0.25}

\item{init.sepsilon}{specifies the initial value for sepsilon. default is 0.1}

\item{init.w.obs.phi}{If TRUE, initialize phi values with observed phi values 
(data from RNAseq, mass spectrometry, ribosome footprinting) Default is FALSE. 
If multiple observed phi values exist for a gene, the geometric mean of these values is used as initial phi.
When using this function, one should remove any genes with 
missing phi values, as these genes will not have an initial phi value.}

\item{init.by.random}{If TRUE, initialize codon-specific parameters randomly. Default is FALSE.}

\item{init.initiation.cost}{FOR FONSE ONLY. Initializes the initiation cost a_1 at this value.}

\item{init.partition.function}{FOR PANSE ONLY. initializes the partition function Z.}

\item{numElongationMixtures}{FOR PANSE ONLY. Number of elongation mixture components. Default is 1.}

\item{phi.prior}{Specifies the prior on the synthesis rate phi. Valid values
are \code{"lognormal"} (default, legacy single-LogNormal behavior) and
\code{"mixture-lognormal"} (mixture of two LogNormals, with one component's
log-location derived from the constraint below). Currently only the ROC
model uses this argument; ignored otherwise.}

\item{phi.prior.constraint}{Anchor for the phi prior. Valid values are
\code{"mean"} (default; anchors E[phi] = 1 via mPhi = -sphi^2/2) and
\code{"median"} (anchors median[phi] = 1 via mPhi = 0). Applies to both
\code{phi.prior = "lognormal"} and \code{phi.prior = "mixture-lognormal"}.
The default (\code{"mean"}) reproduces legacy behavior exactly.}

\item{phi.prior.init}{Optional list with initial values for the mixture-LN
parameters. Named elements \code{p}, \code{mu1}, \code{sigma1}, \code{sigma2}
each numeric (scalar recycled to num.mixtures, or vector of length
num.mixtures). NULL uses defaults from the validated R prototype.}

\item{phi.prior.hyperparams}{Optional named list overriding mixture prior
hyperparameters. Sub-blocks: \code{p = list(alpha=, beta=)} (Beta prior),
\code{mu1 = list(mean=, sd=)} (Normal prior), \code{sigma1 = list(scale=)}
and \code{sigma2 = list(scale=)} (half-Normal scales). Any field omitted
keeps its default. Defaults match the R prototype: Beta(8,2), Normal(0,10),
half-Normal(1), half-Normal(1) truncated above by sigma1.}

\item{phi.mphi}{Optional specification of the phi-prior location (mPhi), built
with \code{\link{constrained}} or \code{\link{fixed}}; e.g.
\code{constrained(statistic = "median", value = 1)}. When supplied it supersedes
the legacy \code{phi.prior.constraint} argument. \code{NULL} (default) uses the
legacy \code{phi.prior.constraint} path. See the phi-prior parameterization
vignette for guidance on the anchor statistic.}

\item{phi.sphi}{Optional specification of sphi (the phi-prior spread), built with
\code{\link{estimated}} or \code{\link{fixed}}; e.g.
\code{estimated(prior = prior_normal(mean = 1.4, sd = 0.125))} or
\code{fixed(value = 1.4)}. When supplied it supersedes the legacy \code{sphi}
argument. \code{NULL} (default) falls back to \code{sphi}.}
}
\value{
parameter Returns an initialized Parameter object.
}
\description{
\code{initializeParameterObject} initializes a new parameter object or reconstructs one from a restart file
}
\details{
\code{initializeParameterObject} checks the values of the arguments 
given to insure the values are valid.

The mixture definition and mixture definition matrix describes how the mutation
and selection categories are set up with respect to the number of mixtures. For
example, if mixture.definition = "allUnique" and numMixtures = 3, a matrix
representation would be \code{matrix(c(1,2,3,1,2,3), ncol=2)}
where each row represents a mixture, the first column represents the mutation
category, and the second column represents the selection category.
Another example would be mixture.definition = "selectionShared" and numMixtures = 4 (
\code{matrix(c(1,2,3,4,1,1,1,1), ncol=2)}).
In this case, the selection category is the same for every mixture. If a matrix
is given, and it is valid, then the mutation/selection relationship will be
defined by the given matrix and the keyword will be ignored. A matrix should only
be given in cases where the keywords would not create the desired matrix.
}
\examples{

genome_file <- system.file("extdata", "genome.fasta", package = "AnaCoDa")
restart_file <- system.file("extdata", "restart_file.rst", package = "AnaCoDa")

genome <- initializeGenomeObject(file = genome_file)

## initialize a new parameter object
sphi_init <- 1
numMixtures <- 1
geneAssignment <- rep(1, length(genome))
parameter <- initializeParameterObject(genome = genome, sphi = sphi_init, 
                                       num.mixtures = numMixtures, 
                                       gene.assignment = geneAssignment, 
                                       mixture.definition = "allUnique")

## re-initialize a parameter object from a restart file. Useful for checkpointing
parameter <- initializeParameterObject(init.with.restart.file = restart_file)

## initialize a parameter object with a custon mixture definition matrix
def.matrix <- matrix(c(1,1,1,2), ncol=2)
geneAssignment <- c(rep(1,floor(length(genome)/2)),rep(2,ceiling(length(genome)/2)))
parameter <- initializeParameterObject(genome = genome, sphi = c(0.5, 2), num.mixtures = 2,
                                       gene.assignment = geneAssignment,
                                       mixture.definition.matrix = def.matrix)

}
