% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{initializeSynthesisRateByGenome}
\alias{initializeSynthesisRateByGenome}
\title{Initialize synthesis rates using SCUO values calcuated from the genome}
\arguments{
\item{genome}{a Genome object}
}
\description{
Initialize synthesis rates using SCUO values calcuated from the genome
}
