% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{initializeSynthesisRateByList}
\alias{initializeSynthesisRateByList}
\title{Initialize synthesis rates with values passed in as a list}
\arguments{
\item{expression}{a list of values to use as initial synthesis rate values. Should be same size as number of genes in genome.}
}
\description{
Initialize synthesis rates with values passed in as a list
}
