% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptiveScheme.R
\name{is.AdaptiveScheme}
\alias{is.AdaptiveScheme}
\title{Test whether an object was constructed by one of the AdaptiveScheme
constructors.}
\usage{
is.AdaptiveScheme(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
Logical.
}
\description{
Test whether an object was constructed by one of the AdaptiveScheme
constructors.
}
