% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterObject.R
\name{loadParameterObject}
\alias{loadParameterObject}
\title{Load Parameter Object}
\usage{
loadParameterObject(files)
}
\arguments{
\item{files}{A list of parameter filenames to be loaded. If multiple files are given, 
the parameter objects will be concatenated in the order provided}
}
\value{
parameter Returns an initialized Parameter object depending on the parameter type given.
}
\description{
\code{loadParameterObject} will call the appropriate followup
call to loadXXXParameterObject based off of the parameter type 
given.
}
\details{
For example, if a ROCParameter is passed, the the loadParameterObject
for the ROCParameter will be called. This allows us to not have an if-else
block in the code - making use of the how R handles these situations.
}

