% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMCMCAlgorithmObject.R
\name{mcmcACF}
\alias{mcmcACF}
\title{Plots ACF for MCMC traces}
\usage{
mcmcACF(mcmc, type = "LogPosterior", samples = 500, lag.max = 40)
}
\arguments{
\item{mcmc}{object of class MCMC}

\item{type}{"LogPosterior" or "LogLikelihood", defaults to "LogPosterior"}

\item{samples}{number of samples to calculate ACF for,
Ex) if samples == 300 and hae 1000 samples total, ACF will be calculated from samples 700 to 1000}

\item{lag.max}{Maximum amount of lag to calculate ACF}
}
\description{
Plots ACF for MCMC traces
}

