% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModelObject.R
\name{plot.Rcpp_FONSEModel}
\alias{plot.Rcpp_FONSEModel}
\title{Plot Model Object}
\usage{
\method{plot}{Rcpp_FONSEModel}(x, genome, parameter, samples = 100,
  mixture = 1, estim.Expression = TRUE, simulated = FALSE, ...)
}
\description{
See \code{plot.Rcpp_ROCModel}
}

