% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotParameterObject.R
\name{plot.Rcpp_ROCParameter}
\alias{plot.Rcpp_ROCParameter}
\title{Plot ROC Parameter}
\usage{
\method{plot}{Rcpp_ROCParameter}(x, what = "Mutation", samples = 100, ...)
}
\arguments{
\item{x}{ROC parameter object.}

\item{what}{Which aspect of the ROC parameter to plot. Default value is
"Mutation".}

\item{samples}{Number of samples to plot using the posterior mean. Default
value is 100.}

\item{...}{Arguments to be passed to methods, such as graphical parameters.}
}
\value{
This function has no return value.
}
\description{
\code{plot} graphs the mutation or selection parameter for a ROC
parameter object for each mixture element.
}
\details{
Graphs are based off the last # samples for the posterior mean.
}

