% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptiveScheme.R
\name{print.AdaptiveScheme}
\alias{print.AdaptiveScheme}
\title{Print an AdaptiveScheme.}
\usage{
\method{print}{AdaptiveScheme}(x, ...)
}
\arguments{
\item{x}{An AdaptiveScheme.}

\item{...}{Unused.}
}
\description{
Print an AdaptiveScheme.
}
