% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi.prior.spec.R
\name{print.phi_spec}
\alias{print.phi_spec}
\title{Print a phi_spec object}
\usage{
\method{print}{phi_spec}(x, ...)
}
\arguments{
\item{x}{A \code{phi_spec} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
\code{x} invisibly.
}
\description{
Print a phi_spec object
}
