% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi.prior.spec.R
\name{print.prior_dist}
\alias{print.prior_dist}
\title{Print a prior_dist object}
\usage{
\method{print}{prior_dist}(x, ...)
}
\arguments{
\item{x}{A \code{prior_dist} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
\code{x} invisibly.
}
\description{
Print a prior_dist object
}
