% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi.prior.spec.R
\name{prior_normal}
\alias{prior_normal}
\title{Normal prior for MCMC parameter estimation}
\usage{
prior_normal(mean = 0, sd = 1)
}
\arguments{
\item{mean}{Prior mean (default 0).}

\item{sd}{Prior standard deviation (default 1). Must be positive.}
}
\value{
A \code{prior_dist} object.
}
\description{
Normal prior for MCMC parameter estimation
}
