% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi.prior.spec.R
\name{prior_uniform}
\alias{prior_uniform}
\title{Uniform prior for MCMC parameter estimation}
\usage{
prior_uniform(low = 0, high = 10)
}
\arguments{
\item{low}{Lower bound (default 0). Must be >= 0 for sphi.}

\item{high}{Upper bound (default 10).}
}
\value{
A \code{prior_dist} object.
}
\description{
Uniform prior for MCMC parameter estimation
}
