% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readPhiValue}
\alias{readPhiValue}
\title{Read synthesis rate values from file. File should be two column file <gene_id,phi> and is expected to have a header row}
\arguments{
\item{filename}{name of file to be read}
}
\description{
Read synthesis rate values from file. File should be two column file <gene_id,phi> and is expected to have a header row
}
