% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setAdaptiveWidth}
\alias{setAdaptiveWidth}
\title{Set sample adaptiveWidth value, which is the number of samples (not iterations) between adapting parameter proposal widths}
\arguments{
\item{_adaptiveWidth}{postive value}
}
\description{
Set sample adaptiveWidth value, which is the number of samples (not iterations) between adapting parameter proposal widths
}
