% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterObject.R
\name{setBaseInfo}
\alias{setBaseInfo}
\title{Set Base Info}
\usage{
setBaseInfo(parameter, files)
}
\arguments{
\item{parameter}{A Parameter object that corresponds to one of the model types.
Valid values are "ROC", "RFP", and "FONSE".}

\item{files}{A set of files that are loaded to initialize the Parameter object.}
}
\value{
parameter Returns an initialized Parameter object.
}
\description{
Set Base Info
}

