% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setIterationsToAdapt}
\alias{setIterationsToAdapt}
\title{setIterationsToAdapt}
\arguments{
\item{iterations}{a postive value}
}
\description{
Method of MCMC class (access via mcmc$<function name>, where mcmc is an object initialized by initializeMCMCObject). Set number of iterations (total iterations = samples * thinning) to allow proposal widths to adapt
}
