% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setRestartFileSettings}
\alias{setRestartFileSettings}
\title{Set restart file output name and frequency prior to running MCMC}
\arguments{
\item{filename}{name of restart file}

\item{interval}{number of samples (ie. iterations * thinning) between writing new restart file}

\item{multiple}{if true, will output a new restart file at each interval (file name will include sample it was written at)}
}
\description{
Set restart file output name and frequency prior to running MCMC
}
