% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setSamples}
\alias{setSamples}
\title{Set number of samples set for MCMCAlgorithm object}
\arguments{
\item{_samples}{postive value}
}
\description{
Set number of samples set for MCMCAlgorithm object
}
