% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setStepsToAdapt}
\alias{setStepsToAdapt}
\title{Set number of iterations (total iterations = samples * thinning) to allow proposal widths to adapt}
\arguments{
\item{steps}{postive value}
}
\description{
Set number of iterations (total iterations = samples * thinning) to allow proposal widths to adapt
}
