% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{shareNSERate}
\alias{shareNSERate}
\title{shareNSERate}
\description{
Method of PANSEParameter class (access via parameter$shareNSERate(), where parameter is initialized with model="PANSE"). Share (pool) the nonsense error rate across all mixture categories.
}
