% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterObject.R
\name{initializeParameterObject}
\alias{initializeParameterObject}
\title{Initialize Parameter}
\usage{
initializeParameterObject(genome = NULL, sphi = NULL, num.mixtures = 1,
  gene.assignment = NULL, initial.expression.values = NULL, model = "ROC",
  split.serine = TRUE, mixture.definition = "allUnique",
  mixture.definition.matrix = NULL, init.with.restart.file = NULL,
  mutation.prior.sd = 0.35, init.csp.variance = 0.0025)
}
\arguments{
\item{genome}{An object of type Genome necessary for the initialization of the Parameter object.
The default value is NULL.}

\item{sphi}{Initial values for sphi. Expected is a vector of length numMixtures.
The default value is NULL.}

\item{num.mixtures}{The number of mixtures elements for the underlying mixture distribution (numMixtures > 0).
The default value is 1.}

\item{gene.assignment}{A vector holding the initial mixture assignment for each gene. 
The vector length has to equal the number of genes in the genome.
Valid values for the vector range from 1 to numMixtures. 
It is possible but not advised to leave a mixture element empty.
The default Value is NULL.}

\item{initial.expression.values}{(Optional) A vector with intial phi values.
The length of the vector has to equal the number of genes in the Genome object.
The default value is NULL.}

\item{model}{Specifies the model used. Valid options are "ROC", "RFP", or "FONSE".
The default model is "ROC".
ROC is described in Gilchrist et al. 2015.
RFP and FONSE are currently unpublished.}

\item{split.serine}{Whether serine should be considered as 
one or two amino acids when running the model.
TRUE and FALSE are the only valid values.
The default value for split.serine is TRUE.}

\item{mixture.definition}{A string describing how each mixture should
be treated with respect to mutation and selection.
Valid values consist of "allUnique", "mutationShared", and "selectionShared".
The default value for mixture.definition is "allUnique".
See details for more information.}

\item{mixture.definition.matrix}{A matrix representation of how
the mutation and selection categories correspond to the mixtures.
The default value for mixture.definition.matrix is NULL.
If provided, the model will use the matrix to initialize the mutation and selection
categories instead of the definition listed directly above.
See details for more information.}

\item{init.with.restart.file}{File name containing information to reinitialize a 
previous Parameter object.
If given, all other arguments will be ignored.
The default value for init.with.restart.file is NULL.}

\item{mutation.prior.sd}{Controlling the standard deviation of the normal 
prior on the mutation parameters}

\item{init.csp.variance}{specifies the initial proposal width for codon specific parameter (default is 0.0025). 
The proposal width adapts during the runtime to reach a taget acceptance rate of ~0.25}
}
\value{
parameter Returns an initialized Parameter object.
}
\description{
\code{initializeParameterObject} will call the appropriate followup
call to writeXXXParameterObject based off of the value of model and init.with.restart.file.
}
\details{
\code{initializeParameterObject} checks the values of the arguments 
given to insure the values are valid.
Additionally, if a restart file is given, no follow up function calls are made - 
a new call is made instead which calls the C++ constructor that only takes a file name.

The mixture definition and mixture definition matrix describe how the mutation
and selection categories are set up with respect to the number of mixtures. For
example, if mixture.definition = "allUnique" and numMixtures = 3, a matrix
representation would be as follows:

1 1

2 2

3 3

where each row represents a mixture, the first column represents the mutation
category, and the second column represents the selection category.
Another example would be mixture.definition = "selectionShared" and numMixtures = 4.

1 1

2 1

3 1

4 1

In this case, the selection category is the same for every mixture. If a matrix
is given, and it is valid, then the mutation/selection relationship will be
defined by the given matrix as opposed to the keyword. A matrix should only
be given in cases where the keywords would not create the desired valid matrix.

If expressionValues is given, then the phi values will be initialized with 
them. If not, we calculate starting phi values by doing an SCUO caculation.
}

