% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi.prior.spec.R
\name{prior_exponential}
\alias{prior_exponential}
\title{Exponential prior for MCMC parameter estimation}
\usage{
prior_exponential(rate = 1)
}
\arguments{
\item{rate}{Rate parameter (default 1). Must be positive.}
}
\value{
A \code{prior_dist} object.
}
\description{
Exponential prior for MCMC parameter estimation
}
\examples{
prior_exponential(rate = 1)
}
