% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi.prior.spec.R
\name{prior_student_t}
\alias{prior_student_t}
\title{Student-t prior for MCMC parameter estimation}
\usage{
prior_student_t(df = 3, mean = 0, sd = 1)
}
\arguments{
\item{df}{Degrees of freedom (default 3). Must be positive.}

\item{mean}{Location (default 0).}

\item{sd}{Scale (default 1). Must be positive.}
}
\value{
A \code{prior_dist} object.
}
\description{
Student-t prior for MCMC parameter estimation
}
